O1: Send spatial_features as float16 from LitServe server
Halves the response payload from ~10MB to ~5MB by returning
bd['spatial_features'].half().numpy() instead of float32.

The edge manager already calls .float().cuda() before running the
Where2comm backbone, so no additional cast is needed on the client side.

Measured improvement on vehicle agent (n=168 ticks):
  http_ms:      291ms → 175ms  (-116ms, -40%)
  total_e2e_ms: 355ms → 234ms  (-121ms, -34%)
  response_kb:  10000 →  5000  (-50%, exact)

The http_ms gap (http - server_total) halved in proportion with the
response size (220ms → 103ms), confirming the bottleneck is TCP download
of the response body, not server processing or request upload.

+++ ecav/ml_manager/litserve_models.py
@@ -211,7 +211,9 @@ def _extract_wf_features(batch):
         b, c, z, y, x_dim = spatial_3d.shape
         bd['spatial_features'] = spatial_3d.view(b, c * z, y, x_dim)
 
-    return {'spatial_features': bd['spatial_features'].cpu().numpy()}
+    # Return as float16 to halve response payload (~10MB → ~5MB).
+    # The edge manager casts back to float32 via .float().cuda() before fusion.
+    return {'spatial_features': bd['spatial_features'].cpu().half().numpy()}
 
 
 if __name__ == "__main__":
